Added NTN-FR2 FDD bands defined in release 18 in Ka-band range

See merge request oai/openairinterface5g!3414

+++ executables/nr-softmodem-common.h
@@ -122,7 +122,7 @@ extern int sync_var;
 
 
 extern uint64_t downlink_frequency[MAX_NUM_CCs][4];
-extern int32_t uplink_frequency_offset[MAX_NUM_CCs][4];
+extern int64_t uplink_frequency_offset[MAX_NUM_CCs][4];
 
 extern int oai_exit;
